TSDK-678 Fix conversation module
