--- daml-topl-broker/src/main/scala/co/topl/broker/processors/SimpleTransactionAlgebra.scala
@@ -1,20 +1,0 @@
-package co.topl.broker.processors
-
-
-
-sealed trait SimpleTransactionAlgebraError {
-
-  def description: String
-}
-
-case class CannotInitializeProtobuf(description: String) extends SimpleTransactionAlgebraError
-
-case class InvalidProtobufFile(description: String) extends SimpleTransactionAlgebraError
-
-case class CannotSerializeProtobufFile(description: String) extends SimpleTransactionAlgebraError
-
-case class NetworkProblem(description: String) extends SimpleTransactionAlgebraError
-
-case class UnexpectedError(description: String) extends SimpleTransactionAlgebraError
-
-case class CreateTxError(description: String) extends SimpleTransactionAlgebraError
