--- daml-topl-broker/src/main/scala/co/topl/broker/processors/package.scala
@@ -1,9 +1,0 @@
-package co.topl.broker
-
-import cats.data.Kleisli
-
-package object processors {
-
-  type DAMLKleisli[F[_], A] = Kleisli[F, DAMLContext, A]
-
-}
